backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
google-generativeai==0.3.2
elevenlabs==0.2.26
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
